move clientCredential test to correct location
